*Updated README, notebooks *Removed circular import *Added comments on loader files * separated cross-compile script *separated utils funcs
