+++ QEfficient/loader/loader.py
@@ -5,6 +5,12 @@
 #
 # -----------------------------------------------------------------------------
 
+"""
+MODEL_TYPE_TO_QEFF_AUTO_MODEL_MAP dictionary defines the mapping between names of the varities of Transformer model defined in 
+QEFF_MODEL_TYPE and the classes that implement the methods i.e.(compile, export etc.) for those types.
+
+QEFFAutoModel provides a common interface for loading the HuggingFace models using either the HF card name of local path of downloaded model.
+"""
 import os
 from typing import Any, Dict, Type
