+++ tests/utils.py
@@ -11,7 +11,7 @@ import shutil
 import unittest
 
 from QEfficient.exporter.export_hf_to_cloud_ai_100 import qualcomm_efficient_converter
-from QEfficient.exporter.export_utils import compile_kv_model_on_cloud_ai_100
+from QEfficient.cross_compile import compile_kv_model_on_cloud_ai_100
 from QEfficient.loader.loader_factory import QEFFAutoModelForCausalLM
 from QEfficient.transformers.transform import transform_lm
 from QEfficient.utils import hf_download, load_hf_tokenizer
